+++ openvmm/openvmm_defs/src/config.rs
@@ -57,7 +57,7 @@ pub struct Config {
     pub chipset_capabilities: VmChipsetCapabilities,
     /// Chipset low MMIO range size (below 4 GiB) for VMOD/PCI0 _CRS.
     /// The address is always allocated dynamically. `0` disables the range.
-    pub chipset_low_mmio_size: u64,
+    pub chipset_low_mmio_size: u32,
     /// Chipset high MMIO range size (above RAM) for VMOD/PCI0 _CRS.
     /// The address is always allocated dynamically. `0` disables the range.
     pub chipset_high_mmio_size: u64,
